requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
fpdf2==2.7.9
pandas==2.2.2
openpyxl==3.1.2
python-multipart==0.0.9
